chore: handle all the supported GUIDs
UEFI SCT test would expect the supported GUID
types to work on both GetInformation() and
SetInformation().

For the original supported GUIDs which are not
handled in GetInformation(), returning EFI_SUCCESS
and one byte info block.

For the original supported GUIDs which are not
handled in SetInformation(), returning EFI_SUCCESS
directly.

Signed-off-by: Simon Wang <simowang@nvidia.com>
Reviewed-by: Blake Tsai <blaket@nvidia.com>
Reviewed-by: Dat Mach <dmach@nvidia.com>
Reviewed-by: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Nickle Wang <nicklew@nvidia.com>